# feat(bigquery): support reading parquet, json and csv files