This PR adds support for reading Parquet, JSON and CSV files directly into BigQuery. Local paths, local globs, GCS paths, and GCS globs are all supported. Depends on #7545.

Loading local globs with and without `concurrent.futures`:

### Setup

```
cloud in 🌐 falcon in ~ via 🐍
❯ ls ~/data/trip-data
Permissions Size User  Date Modified Name
.rw-r--r--  151M cloud 26 Jun 11:34  yellow_tripdata_2016-01.parquet
.rw-r--r--  158M cloud 26 Jun 11:34  yellow_tripdata_2016-02.parquet
```

### WITH `concurrent.futures`

```
In [1]: from ibis.interactive import *

In [2]: con = ibis.bigquery.connect()

In [3]: ibis.set_backend(con)

In [4]: %time t = ibis.read_parquet("/home/cloud/data/trip-data/*.parquet")
CPU times: user 254 ms, sys: 191 ms, total: 445 ms
Wall time: 51 s
```

### WITHOUT `concurrent.futures`

```
In [1]: from ibis.interactive import *

In [2]: con = ibis.bigquery.connect()

In [3]: ibis.set_backend(con)

In [4]: %time t = ibis.read_parquet("/home/cloud/data/trip-data/*.parquet")
CPU times: user 213 ms, sys: 216 ms, total: 429 ms
Wall time: 1min 24s
```

So there's a nice latency decrease by running the load job for each file concurrently (this is using threads).