@cpcloud 2023-11-15
BQ tests passing:

```
…/ibis on  bigquery-read-files is 📦 v7.0.0 via 🐍 v3.10.13 via ❄️  impure (ibis-3.10.13-env) took 3m49s
❯ pytest -m bigquery --snapshot-update -q -n auto --dist loadgroup
bringing up nodes...
ssssssssssssssssssssss.x.....x......x..x.x....x............xxx.............xx......xx..xx.xx....x.x.x........x.........xx..x....................x........................xx...x...x.......................... [ 13%]
............................................x..xx..........................x.x.....x..x..xx.x..xx.x..x...........x......x..x.....x......................x...xx.x.x.x.......x..x.x.x.x.........xx............. [ 26%]
x......x.......x...................x.x...x........x......x.x.x..xxxxxx.xx.xxxxxxxxxxxxxxx.x...xxxxxx.xxxx..xxxxxx........x..xx.x...x.x..x.x..x...xxx.x.......s.......xx.....x.........x.............s...x.... [ 39%]
.......................................................s......x................................................x.............................x...............xx...x.......x...x....x.........x.x.x..x........ [ 52%]
...x...x...x.....xx.xx.x......x.x....x...............x...x.....x...x........xx.xx.x....x.xxxx..x.x.x...x.x.x...x.s.....x.x...x...........x..x......x..xx..x...xxx.xx..xx.x.x..x...x.x.x........xxx.x......... [ 65%]
.....xxx.x....xx.............x.x....xx...x.x.x.xxxx.....x.....x...xxxx.x........x...xx...x.x.x............x.x...x....xx......................x...................................xx..xxx...x..........x..x... [ 78%]
............................................................................................................................................................................................................. [ 91%]
...........................................x......................................................x........s..........x.........                                                                              [100%]
1290 passed, 27 skipped, 246 xfailed in 369.33s (0:06:09)
```